person service

+++ src/main/java/com/vmo/apartment_manager/entity/Person.java
@@ -16,6 +16,6 @@ public class Person extends BaseEntity{
   private String cin;
   private Boolean gender;
   private String carrer;
-  private Integer idParent;
+  private Long idParent;
 
 }
